Fixed marker model lat/lng data type, added test dir w/starter spec files

+++ db/models/marker.js
@@ -13,11 +13,11 @@ const Marker = db.define('marker', {
     allowNull: false,
   },
   latitude: {
-    type: Sequelize.FLOAT(15),
+    type: Sequelize.FLOAT,
     allowNull: false,
   },
   longitude: {
-    type: Sequelize.FLOAT(15),
+    type: Sequelize.FLOAT,
     allowNull: false,
   },
   notes: {
